Make the dynamic send starttimes nomatter what. Include the boolean. Solve it on arduino, I will save loads of time.
Do this by setting "getStartTime" to allways true on javascript.
And instead of changeing the boolean when length == 0, I will instead
set starttime to 0 millisecoonds.
this will make backend so mutch eazier... should i include shit_times?
Sure, but just ignore it in the domain.....
or add null when accessing.

Need more design, but almost done and fuck databases!